JPetStoreDemo
JPetStore Demo
The JPetStore Demo is an online pet store. Like most e-stores, you can browse and search the product catalog, 
choose items to add to a shopping cart, and add the items in the shopping cart. 
You can perform many of these actions without registering with or logging into the application. However, 
before you can order items you must log in (sign in) to the application. In order to sign in,
you must have an account with the application, which is created when you register (sign up) with the application.